pyproject.toml
[tool.poetry]
name = "acme-nginx"
version = "0.3.0"
description = "Python library/program to create LetsEncrypt SSL certificates"
authors = ["Konstantin Shcherban <insider@sudoers.dev>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.6"
pyOpenSSL = "^20.0"
pycrypto = "^2.6"
boto3 = "~1.17"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
